requirements.txt
mss>=9.0.1
pillow>=10.4.0
requests>=2.31.0
plotly>=5.20.0
pandas>=2.2.1
pyyaml>=6.0.1
rumps>=0.4.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.5
python-socketio>=5.10.0
